fix(admin): extract computeServiceStatuses to server-side lib to prevent client-bundle TypeError
Health page was importing computeServiceStatuses from ServiceStatusGrid.tsx ("use client"),
causing `m is not a function` at runtime because the server component attempted to use a
function defined inside a client-boundary module.

Fix: extract MONITORED_SERVICES + computeServiceStatuses + ServiceSummary/ServiceStatus types
into lib/platform/service-health/status.ts (no server-only guard — pure utility, safe on both
sides). ServiceStatusGrid now re-exports MONITORED_SERVICES from the lib; BillingIssueDialog
and health page.tsx both import directly from the lib.
